scrapers/pyproject.toml
[project]
name = "carnaval-scrapers"
version = "0.1.0"
description = "Web scrapers para recolectar datos del Carnaval de Barranquilla"
requires-python = ">=3.9"
dependencies = [
    "scrapy>=2.11",
    "beautifulsoup4>=4.12",
    "requests>=2.31",
    "playwright>=1.40",
    "lxml>=5.1",
    "python-dotenv>=1.0",
    "supabase>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "black>=24.0",
    "ruff>=0.2",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.black]
line-length = 100
target-version = ["py39"]
